Create before_after_main.c
With GCC family of C compilers, we can mark some functions to execute before and after main(). So some startup code can be executed before main() starts, and some cleanup code can be executed after main() ends
This code is taken from https://www.geeksforgeeks.org/functions-that-are-executed-before-and-after-main-in-c/, All rights of this code belongs to www.geeksforgeeks.org.